Address sanitizer flags

+++ Makefile
@@ -1,13 +1,15 @@
 INCLUDE_DIR=./include/
 EXEC=advcalc
 CC=gcc
-CFLAGS=-I $(INCLUDE_DIR) -o $(EXEC) -Wall -Wextra -Werror -pedantic 
+CFLAGS=-I $(INCLUDE_DIR) -o $(EXEC) -Wall -Wextra -Werror -pedantic #-fsanitize=address -g
 
 SRC=$(filter-out ./src/test.c, $(wildcard ./src/*.c))
 
-clean :
-	rm $(EXEC)
+default: clean advcalc
 
+clean :
+	if [ -f $(EXEC) ] ; then rm $(EXEC) ; fi
+	
 advcalc:  
 	$(CC) $(SRC) $(CFLAGS)
